Move batch method tests to their own file
